backend/requirements.txt
Flask==3.1.0
opencv-python==4.10.0.84
tensorflow==2.17.0
numpy==1.26.4
face_recognition==1.3.0
scikit-learn==1.5.2
flask-cors-5.0.0
flask-socketio-5.4.1
gevent==24.11.1
gunicorn==23.0.0